+++ src/elements/Poll.php
@@ -13,14 +13,13 @@ use craft\elements\db\ElementQueryInterface;
 use craft\elements\db\EntryQuery;
 use craft\elements\Entry;
 use craft\helpers\UrlHelper;
-use twentyfourhoursmedia\commentswork\elements\db\CommentQuery;
 use twentyfourhoursmedia\poll\elements\actions\CreateReport;
 use twentyfourhoursmedia\poll\elements\db\PollQuery;
 
 class Poll extends Entry
 {
 
-    protected static function defineActions(string $source = null): array
+    protected static function defineActions(string $source): array
     {
         $actions = [];
         // Edit
@@ -53,14 +52,14 @@ class Poll extends Entry
 
     /**
      * Creates an [[ElementQueryInterface]] instance for query purpose.
-     * @return CommentQuery.
+     * @return PollQuery
      */
     public static function find(): EntryQuery
     {
         return new PollQuery(static::class);
     }
 
-    protected static function defineSources(string $context = null): array
+    protected static function defineSources(string $context): array
     {
         $sources = [];